+++ WinQuake/pr_cmds.c
@@ -572,8 +572,8 @@ void PF_sound (void)
 	if (volume < 0 || volume > 255)
 		Sys_Error ("SV_StartSound: volume = %i", volume);
 
-	if (attenuation < 0 || attenuation > 4)
-		Sys_Error ("SV_StartSound: attenuation = %f", attenuation);
+	if (attenuation < 0.0F || attenuation > 4.0F)
+		Sys_Error ("SV_StartSound: attenuation = %f", (double)attenuation);
 
 	if (channel < 0 || channel > 7)
 		Sys_Error ("SV_StartSound: channel = %i", channel);
@@ -928,9 +928,9 @@ void PF_ftos (void)
 	v = G_FLOAT(OFS_PARM0);
 	
 	if (v == (int)v)
-		sprintf (pr_string_temp, "%d",(int)v);
+		sprintf (pr_string_temp, "%d", (int)v);
 	else
-		sprintf (pr_string_temp, "%5.1f",v);
+		sprintf (pr_string_temp, "%5.1f", (double)v);
 	G_INT(OFS_RETURN) = pr_string_temp - pr_strings;
 }
 
@@ -943,7 +943,8 @@ void PF_fabs (void)
 
 void PF_vtos (void)
 {
-	sprintf (pr_string_temp, "'%5.1f %5.1f %5.1f'", G_VECTOR(OFS_PARM0)[0], G_VECTOR(OFS_PARM0)[1], G_VECTOR(OFS_PARM0)[2]);
+	sprintf (pr_string_temp, "'%5.1f %5.1f %5.1f'", (double)G_VECTOR(OFS_PARM0)[0],
+			 (double)G_VECTOR(OFS_PARM0)[1], (double)G_VECTOR(OFS_PARM0)[2]);
 	G_INT(OFS_RETURN) = pr_string_temp - pr_strings;
 }
